% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dendrometeR.R, R/min_resolution.R
\name{as_dendrometeR}
\alias{as_dendrometeR}
\title{Convert data.frame/tibble to dendrometeR format}
\usage{
as_dendrometeR(data, timestamp_column = 1)

min_resolution(data, timestamp_column = 1)
}
\value{
A \code{data.frame}

A \code{data.frame}
}
\description{
convert to the dendrometeR format

return minimal resolution of the dendrometer series
}
\author{
Marko Smiljanic
}
