% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_data.R
\name{assign_data}
\alias{assign_data}
\title{Assign sensor data to the data collection}
\usage{
assign_data(data, collection, index = timeindex,
            by_vector = c('TIMESTAMP' = 'TIMESTAMP'),
            column = 'Sensor', sensorID = column)
}
\value{
A \code{data.frame}
}
\description{
transform series timestamps to the uniform resolution
}
\examples{

}
\author{
Marko Smiljanic
}
