% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_jump.R
\name{auto_jump}
\alias{auto_jump}
\title{Detect jumps in the signal with the simple algorithm}
\usage{
auto_jump(data, threshold, timestamp_column = 1, name = 'Sensor')
}
\value{
A \code{data.frame}
}
\description{
detect jump corrections
}
\author{
Marko Smiljanic
}
