% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_years.R
\name{compare_years}
\alias{compare_years}
\title{Compare yearly seasonality}
\usage{
compare_years(data, timestamp_column = 1)
}
\value{
A \code{data.frame}
}
\description{
compares all of the years
}
\author{
Marko Smiljanic
}
