% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELU.R
\docType{data}
\name{ELU}
\alias{ELU}
\title{Pre-processed dendrometer data from Eldena, Germany}
\format{
A \code{data.frame} with a timestamp (\code{\%Y-\%m-\%d \%H:\%M:\%S} format) as row names, and the two dendrometer series in the first and second column.
}
\usage{
data(ELU)
}
\description{
This dataset presents a pre-processed version of \code{\link{dmEDraw}}, using code provided in the \code{Import dendrometer data} vignette.
}
\keyword{datasets}
