% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resolution.R
\name{get_resolution}
\alias{get_resolution}
\title{Get possible resolutions and their frequencies}
\usage{
get_resolution(data, timestamp_column = 1)
}
\value{
A \code{data.frame}
}
\description{
extract temporal resolution frequencies
}
\examples{

}
\author{
Marko Smiljanic
}
