% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_sensor.R
\name{glimpse_sensor}
\alias{glimpse_sensor}
\title{Glimpse sensor first few lines}
\usage{
glimpse_sensor(filenames, n=10, ...)
}
\value{
A \code{list} containing firt few lines of the sensor filenames
}
\description{
transform series timestamps to the uniform resolution
}
\examples{
# extract filepath of the example dataset
file_path <- system.file("extdata", 'Vilm-3_Sensors.dat', package = "dendwrang", mustWork = TRUE)
glimpse_sensor(file_path)

}
