% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_correction.R
\name{jump_correction}
\alias{jump_correction}
\title{Correct signal jumps due to base shifts}
\usage{
jump_correction(data, pre, post, jump, offset = 0, name = 'Sensor')
}
\value{
A \code{data.frame}
}
\description{
setting jump corrections
}
\examples{

}
\author{
Marko Smiljanic
}
