% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumper.R
\name{jumper}
\alias{jumper}
\title{Recursive jump correction}
\usage{
jumper(data, timestamp_column = 1, name = 'Sensor', ...)
}
\value{
A \code{data.frame}
}
\description{
auto correct jump and outlier corrections
}
\author{
Marko Smiljanic
}
