% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_resolution.R
\name{min_resolution}
\alias{min_resolution}
\title{Return minimal resolution}
\usage{
min_resolution(data, timestamp_column = 1)
}
\value{
A \code{data.frame}
}
\description{
return minimal resolution of the dendrometer series
}
\author{
Marko Smiljanic
}
