% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_set.R
\name{na_set}
\alias{na_set}
\title{Remove suspect data}
\usage{
na_set(data, seq, column = 'Sensor', tz = 'UTC', seq_override = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}}

\item{seq}{either a character string or a \code{POSIXct} vector.}

\item{column}{name(s)}

\item{tz}{A \code{character} indicating time zone parameter forwarded to \code{as.POSIXct} function}

\item{seq_override}{a boolean indicating whether the \code{seq} parameter should be treated as a sequence no matter the length}
}
\value{
original \code{data.frame} or \code{tibble} with the \code{NA} in data
}
\description{
function sets single suspect data point or data sequence to \code{NA}
}
\details{

}
\examples{

}
\author{
Marko Smiljanic
}
