% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize the series}
\usage{
normalize(data, timestamp_column = 1)
}
\value{
A \code{data.frame}
}
\description{
compares all of the years
}
\author{
Marko Smiljanic
}
