% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensor.R
\name{plot_sensor}
\alias{plot_sensor}
\title{plot(inspect) sensor}
\usage{
plot_sensor(data, plotly = FALSE, line = TRUE)
}
\value{
A \code{data.frame}
}
\description{
transform series timestamps to the uniform resolution
}
\author{
Marko Smiljanic
}
