% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sensor.R
\name{read_sensor}
\alias{read_sensor}
\title{Read entire sensor}
\usage{
read_sensor(filenames, timestamp_columns, sensor_columns,
            data_rows, header_rows, col_names = NULL, ...)
}
\value{
A \code{data.frame}
}
\description{
reading sensor data
}
\examples{

}
\author{
Marko Smiljanic
}
