% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Assign sensor data to the data collection}
\usage{
report(timeindex, cleaning_script, plot_name, filename, scales = 'fixed')
}
\value{
A \code{data.frame}
}
\description{
transform series timestamps to the uniform resolution
}
\author{
Marko Smiljanic
}
