% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_plot.R
\name{report_plot}
\alias{report_plot}
\title{plot collection}
\usage{
report_plot(dat=out, facets = F, scales = 'fixed')
}
\value{
A \code{data.frame}
}
\description{
transform series timestamps to the uniform resolution
}
\examples{

}
\author{
Marko Smiljanic
}
