% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_sample.R
\name{sub_sample}
\alias{sub_sample}
\title{Modifying series resolutions}
\usage{
sub_sample(data, timestamp_column = 1, resolution = 1800, start = NULL, end = NULL)
}
\value{
A \code{data.frame}
}
\description{
transform series timestamps to the uniform resolution
}
\author{
Marko Smiljanic
}
